refactor: useUserProfile 및 userApi 참조 주소변경 및 이전 파일 삭제
