+++ src/pages/user-profile/ui/UserProfilePage.tsx
@@ -3,7 +3,6 @@ import { styled as muiStyled } from "@mui/material/styles";
 import type { JSX } from "react";
 import { useState, useEffect } from "react";
 
-import { useUserProfile } from "@features/auth/hooks/useUserProfile";
 import { removeProjectsFromUser } from "@features/projects/api/projectsApi";
 
 import { useProjectsByIds } from "@entities/projects/hook/useProjectsByIds";
@@ -11,6 +10,7 @@ import UserProfileCard from "@entities/user/ui/user-profile/UserProfileCard";
 import UserProfileHeader from "@entities/user/ui/user-profile/UserProfileHeader";
 import UserProfileProjectList from "@entities/user/ui/user-profile/UserProfileProjectList";
 
+import { useUserProfile } from "@shared/queries/useUserProfile";
 import { useAuthStore } from "@shared/stores/authStore";
 import { useProjectStore } from "@shared/stores/projectStore";
 import LoadingSpinner from "@shared/ui/loading-spinner/LoadingSpinner";
